inference completed
