+++ app/Graph/build.hpp
@@ -1,3 +1,17 @@
-#pragma once
 #include <iostream>
+#include <stdexcept>
+#include <variant>
+#include <vector>
 #include <opencv2/opencv.hpp>
+#include "Weights_Reader/reader_weights.hpp"
+#include "graph/graph.hpp"
+#include "layers/ConvLayer.hpp"
+#include "layers/DropOutLayer.hpp"
+#include "layers/EWLayer.hpp"
+#include "layers/FCLayer.hpp"
+#include "layers/FlattenLayer.hpp"
+#include "layers/InputLayer.hpp"
+#include "layers/OutputLayer.hpp"
+#include "layers/PoolingLayer.hpp"
+
+void build_graph(Tensor input, Tensor output);
